@@ -17,9 +17,10 @@ Gem::Specification.new do |s|
   s.files = Dir["{app,config,db,lib}/**/*", "MIT-LICENSE", "Rakefile", "README.rdoc"]
   s.test_files = Dir["spec/**/*"]
 
-  s.add_dependency 'rails', '~> 6.0.6.1'
+  s.add_dependency 'rails', '~> 6.1.7.6'
 
   s.add_dependency 'mongoid', '~> 7.5.4'
+  s.add_dependency 'bson', '~> 4.3'
 
   s.add_dependency "aasm",                      "~> 4.8"
   s.add_dependency 'active_model_serializers',  '~> 0.10'
