@@ -16,7 +16,7 @@ Gem::Specification.new do |s|
 
   s.files = Dir["{app,config,db,lib}/**/*", "MIT-LICENSE", "Rakefile", "README.rdoc"]
 
-  s.add_dependency 'rails', '~> 6.0.6.1'
+  s.add_dependency 'rails', '~> 6.1.7.6'
   s.add_dependency "slim", "~> 3.0.8"
   s.add_dependency 'mongoid', '~> 7.5.4'
   s.add_dependency "aasm", "~> 4.8.0"
