@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
 class UserMailer < ApplicationMailer
-  ### add_template_helper makes the view helper methods available in the Mailer templates. It does NOT make the methods available in the Mailer itself
-  ### Thus we have to use Include in addition to add_template_helper
-  add_template_helper Config::AcaHelper
-  add_template_helper Config::SiteHelper
-  add_template_helper Config::ContactCenterHelper
-  add_template_helper ::L10nHelper
+  ### helper makes the view helper methods available in the Mailer templates. It does NOT make the methods available in the Mailer itself
+  ### Thus we have to use Include in addition to helper
+  helper Config::AcaHelper
+  helper Config::SiteHelper
+  helper Config::ContactCenterHelper
+  helper ::L10nHelper
   include Config::AcaHelper
   include Config::SiteHelper
   include Config::ContactCenterHelper
