@@ -188,7 +188,7 @@ module BenefitSponsors # rubocop:disable Metrics/ModuleLength
         end
 
         it "should render js template" do
-          expect(response.content_type).to eq "text/javascript"
+          expect(response.content_type).to eq "text/javascript; charset=utf-8"
         end
 
         it "should be a success" do
@@ -217,7 +217,7 @@ module BenefitSponsors # rubocop:disable Metrics/ModuleLength
         end
 
         it 'should render js template' do
-          expect(response.content_type).to eq "text/javascript"
+          expect(response.content_type).to eq "text/javascript; charset=utf-8"
         end
 
         it 'should not create a new staff role' do
