@@ -943,11 +943,14 @@ module FinancialAssistance
       on_new_determination = ::Operations::Individual::OnNewDetermination.new.call({family: self.family, year: self.effective_date.year})
       if on_new_determination.success?
         Rails.logger.info { "Successfully created new enrollment on_new_determination: #{self.hbx_id}" }
+        true
       else
         Rails.logger.error { "Failed while creating enrollment on_new_determination: #{self.hbx_id}, Failure Message: #{on_new_determination.failure}" }
+        false
       end
     rescue StandardError => e
-      Rails.logger.error { "Error while creating enrollment on_new_determination: #{self.hbx_id}, Error: #{e.message}" }
+      Rails.logger.error("Error while creating enrollment on_new_determination: #{self.hbx_id}, Error: #{e.message}")
+      false
     end
 
     def retro_application
