@@ -410,7 +410,8 @@ RSpec.describe EnrollmentPeriod::SpecialEnrollment, :type => :model do
       census_employee.update_attributes!(employee_role_id: employee_role.id)
       employee_role.update_attributes!(census_employee_id: census_employee.id)
       initial_application.update_attributes!(aasm_state: 'draft')
-      expect(shop_sep.send(:next_poss_effective_date_within_range)).to eq ["No eligible plan years present"]
+      result = shop_sep.send(:next_poss_effective_date_within_range)
+      expect(result.message).to eq "No eligible plan years present"
     end
   end
 end
