@@ -9,7 +9,7 @@ RSpec.describe "insured/consumer_roles/immigration_document_options.js.erb" do
       allow(person).to receive(:consumer_role).and_return consumer_role
       assign(:target, person)
       assign(:vlp_doc_target, "naturalization_cert_container")
-      render file: "insured/consumer_roles/immigration_document_options.js.erb"
+      render template: "insured/consumer_roles/immigration_document_options.js.erb"
     end
 
     it "should have form_for" do
@@ -22,7 +22,7 @@ RSpec.describe "insured/consumer_roles/immigration_document_options.js.erb" do
     before :each do
       assign(:target, nil)
       assign(:vlp_doc_target, "naturalization_cert_container")
-      render file: "insured/consumer_roles/immigration_document_options.js.erb"
+      render template: "insured/consumer_roles/immigration_document_options.js.erb"
     end
 
     it "should not have form_for" do
