@@ -30,7 +30,7 @@ if ExchangeTestingConfigurationHelper.individual_market_is_enabled?
     before :each do
       catastrophic_plan_notice.append_hbe
       catastrophic_plan_notice.build
-      render file: 'notices/ivl/final_catastrophic_plan_letter', locals: local_data
+      render template: 'notices/ivl/final_catastrophic_plan_letter', locals: local_data
     end
 
     context 'for matching text' do
